Added jump, but needs debug
I added the jump feature (mostly my own code, unfortunately). It activates whenever you press the spacebar, however, there is a major issue with it. There seems to be some kind of damping?? Because the first time you press spacebar, it reaches the intended height, but the next jump and the series of jumps after that do not.

Minor issues:

1) Small issue, but can be annoying: I had to add all of my code to moveForward.cs. It's not a big deal, I'll clean it up later.
2) Slightly bigger issue: You can jump in the air as well. I should be able to add a boolean later today that fixes the issue.