the stupid solution to the death code
alright, so this does in fact work now (u can playtest if u want, it's good). every spike kills the cube. however, it's a shamefully bad solution, if you can't tell from the code, just look at the hierarchy bar and tell me if it's a big issue. we also want to add the death animation in RespawnAndKillMe.cs.